mcp-a2a/pyproject.toml
[project]
name = "mcp-a2a"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-samples",
    "click>=8.1.8",
    "google-a2a",
    "langchain-community>=0.3.22",
    "langchain-mcp-adapters>=0.0.9",
    "langchain-openai>=0.3.14",
    "langgraph>=0.3.34",
    "langgraph-sdk>=0.1.63",
    "pypdf>=5.4.0",
    "python-dotenv",
]

[project.scripts]
mcp-a2a = "mcp_a2a:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
a2a-samples = { git = "https://github.com/google/A2A", subdirectory = "samples/python" }
google-a2a = { git = "https://github.com/djsamseng/A2A", subdirectory = "samples/python", branch = "prefixPythonPackage" }
